Implementation of issue #9088
Various changes relating to the addition of a new config.json file sitting
at the root of the 'course' folder.

The purpose of the config.json file is to store attributes relating to the
actual configuration of the course, for starters this stores attributes
such as _defaultLanguage, which lets app.js know which folder to look in
for the index.html.

Also included are attributes for screenSize (breakpoints) which are read
from the new theme.json.

During initialisation these settings are all added to Adapt.config.

A new Grunt task has been added to append the theme's theme.json attribute
with the course's config.json.