Fixes issue #160
When a GMCQ component only had one possible option and the 'feedback'
attribute had not been set on item, no feedback was getting displayed to the
user if they selected the wrong answer.

This change will select any specific item feedback (if available) and fall
back to the component feedback.