Amends for completion checks
Added _isInteractionsComplete at a component level so we can check each
time the components in a block are completed. Previously it would only
inform you the first time they were completed.

Also amended how the models listen to their children as previously
anytime one of the specific change attributes was changed, it would
check each model in the collection rather than only those it has a
relationship with.