"Required to complete" feature
This adds ability to specify which components should be completed to mark entire block (and other components) as completed. Block model needs to specify _requireCompletionOf property which can have following values: 'all', 'any' or valid component id that is child of this block. Below are few examples of use. When _requireCompletionOf is missing or empty it defaults to all.

{
    "_id": "b-03",
    "_parentId": "a-02",
    "_type": "block",
    "_requireCompletionOf": "any"
},
{
    "_id": "b-03",
    "_parentId": "a-02",
    "_type": "block",
    "_requireCompletionOf": ["all"]
},
{
    "_id": "b-03",
    "_parentId": "a-02",
    "_type": "block",
    "_requireCompletionOf": "c-05"
},
{
    "_id": "b-03",
    "_parentId": "a-02",
    "_type": "block",
    "_requireCompletionOf": ["c-05","c-10"]
},