Resolve #2439, adds a timeout to Adapt.wait
To prevent an unhandled expection or a call to begin() not having a corresponding end() from blocking rendering of the Adapt content, a 7-second timeout is now added on every begin.  If the time has elapsed since the last begin() the outstanding calls to Adapt.wait.end() will be triggered.

Also added a _.once() call to the for() callback and replaced double quote marks in strings with single quotes.